+++ proxy-server/APIConstants.js
@@ -9,6 +9,7 @@ export const LOW_TO_HIGH = "lowToHigh";
 export const GLOBAL_VARIABLES = "globalVariables";
 export const ERA_0 = "demo-era";
 export const ERA_1 = "era1";
+export const TRIFECTA = "trifecta";
 
 export const NUMBER_OF_TEAMS = 16;
 export const STARTING_YEAR_SUPPLEMENTAL_DRAFT_PICKS = 2025;