+++ src/screens/dynasty/transactions-history/TransactionsHistory.js
@@ -1,5 +1,4 @@
 import React, { useCallback, useEffect, useMemo, useState } from "react";
-import { returnMongoCollection } from "database-management";
 import { capitalize } from "lodash";
 import { useParams } from "react-router-dom";
 import { useSelector } from "react-redux";
@@ -8,12 +7,13 @@ import * as T from "styles/StandardScreen.styles";
 import * as G from "styles/shared";
 import { TransactionsHistoryTable } from "components/table/TransactionsHistoryTable";
 import { FOOTBALL } from "Constants";
+import { api } from "utils/api";
 
 export const TransactionsHistory = () => {
   const { era, sport, year } = useParams();
   const isReady = useSelector((state) => state?.currentVariables?.isReady);
   const { inSeasonLeagues } = useSelector(
-    (state) => state?.currentVariables?.seasonVariables?.dynasty
+    (state) => state?.currentVariables?.seasonVariables?.dynasty,
   );
   const [transactions, setTransactions] = useState([]);
   const [gmsArray, setGmsArray] = useState([]);
@@ -21,11 +21,7 @@ export const TransactionsHistory = () => {
 
   const getAndSetGmsArray = useCallback(async () => {
     // get list of gm abbreviations
-    const gmCollection = await returnMongoCollection("gms", era);
-    const gmData = await gmCollection.find(
-      {},
-      { projection: { abbreviation: 1, name: 1 } }
-    );
+    const gmData = await api.get(`/gms/${era}`);
     const gmNamesArray = gmData.map((gm) => `${gm.name} (${gm.abbreviation})`);
     setGmsArray(gmNamesArray);
   }, [setGmsArray, era]);
@@ -33,16 +29,10 @@ export const TransactionsHistory = () => {
   useEffect(() => {
     const display = async () => {
       if (isReady) {
-        const sportYear = `${sport}${year}`;
-        const transactionsHistoryCollection = await returnMongoCollection(
-          "transactionsHistory",
-          era
-        );
-        const data = await transactionsHistoryCollection.find({ sportYear });
-        const object = data?.[0] ?? {};
+        const object = await api.get(`/transactions/${sport}/${year}`);
         const { lastScraped: lastScrapedString, transactions } = object;
         console.log(
-          `${year} ${sport} Transactions last scraped (Local): ${lastScrapedString}`
+          `${year} ${sport} Transactions last scraped (Local): ${lastScrapedString}`,
         );
         getAndSetGmsArray();
         setTransactions(transactions);
@@ -51,7 +41,7 @@ export const TransactionsHistory = () => {
           const lastScrapedIndex = lastScrapedString.indexOf(",");
           const lastScrapedDay = lastScrapedString.substring(
             0,
-            lastScrapedIndex
+            lastScrapedIndex,
           );
           setLastScrapedDay(lastScrapedDay);
         }