+++ src/screens/trifecta/standings/FootballStandings.js
@@ -22,14 +22,14 @@ import {
 } from "./FootballColumns";
 import { Table } from "components/table/Table";
 import { isYear1AfterYear2, isYear1BeforeYear2 } from "utils/years";
-import { returnMongoCollection } from "database-management";
 import { insertIntoArray } from "utils/arrays";
 import { assignRankPoints } from "utils/standings";
+import { api } from "utils/api";
 
 export const FootballStandings = () => {
   const { year } = useParams();
   const { currentYear, isFootballStarted, isFootballInSeason } = useSelector(
-    (state) => state?.currentVariables?.seasonVariables?.trifecta
+    (state) => state?.currentVariables?.seasonVariables?.trifecta,
   );
   const isReady = useSelector((state) => state?.currentVariables?.isReady);
   const ownerNamesMapping = useSelector((state) => state?.names?.ownerNames);
@@ -47,11 +47,11 @@ export const FootballStandings = () => {
       const display = (
         trifectaStandings,
         h2hStandings,
-        top5Bottom5Standings
+        top5Bottom5Standings,
       ) => {
         if (top5Bottom5Standings) {
           const weeksLength = Object.keys(top5Bottom5Standings?.[0]).filter(
-            (key) => key.includes("week")
+            (key) => key.includes("week"),
           ).length;
           setNumberOfWeeks(weeksLength);
         }
@@ -60,68 +60,9 @@ export const FootballStandings = () => {
         setTop5Bottom5Standings(top5Bottom5Standings);
       };
 
-      // scrape, then display, then save to mongodb with new last scraped
-      const scrape = async (collection) => {
-        if (Object.keys(ownerNamesMapping).length > 0) {
-          const ownerIdsCollection = await returnMongoCollection("ownerIds");
-          const ownerIdsOwnerNames = await ownerIdsCollection.find({});
-          const { h2hScrape } = await standingsScraper(year);
-          const h2hStandingsWithoutNames = await h2hScrapeToStandings(
-            h2hScrape
-          );
-          const h2hStandings = await addOwnerNames(
-            ownerIdsOwnerNames,
-            h2hStandingsWithoutNames
-          );
-
-          const top5Bottom5Collection = await returnMongoCollection(
-            "footballTop5Bottom5Totals"
-          );
-          const top5Bottom5Data = await top5Bottom5Collection.find({ year });
-          const ownerNamesUnderscoredObject = await returnOwnerNamesUnderscored(
-            h2hStandings
-          );
-          const unrankedTop5Bottom5Standings =
-            await calculateTop5Bottom5Standings(
-              top5Bottom5Data[0]["top5Bottom5TotalsObject"],
-              ownerNamesUnderscoredObject
-            );
-          const top5Bottom5Standings = await assignRankPoints(
-            unrankedTop5Bottom5Standings,
-            "winPer",
-            "highToLow",
-            "top5Bottom5TrifectaPoints",
-            10,
-            1
-          );
-
-          const trifectaStandings = await compileTrifectaStandings(
-            h2hStandings,
-            top5Bottom5Standings,
-            ownerNamesMapping
-          );
-
-          display(trifectaStandings, h2hStandings, top5Bottom5Standings);
-
-          // delete, then save to mongodb
-          console.log("Delete, then save to mongodb");
-          await collection.deleteMany({ year });
-          await collection.insertOne({
-            year,
-            lastScraped: format(new Date(), "MM/dd/yy hh:mm a"),
-            trifectaStandings,
-            h2hStandings,
-            top5Bottom5Standings,
-          });
-        }
-      };
-
       // check if need to scrape or just display
       const check = async () => {
-        const collection = await returnMongoCollection("footballStandings");
-        const data = await collection.find({ year });
-        const object = data[0] ?? {};
-        const lastScrapedString = object?.lastScraped;
+        const object = await api.get(`/trifecta/standings/football/${year}`);
         const {
           trifectaStandings,
           h2hStandings,
@@ -129,35 +70,10 @@ export const FootballStandings = () => {
           footballStandings,
         } = object;
 
-        // if not current year or not started or in season, then just display, do not scrape
-        if (!isFootballStarted || !isFootballInSeason || year !== currentYear) {
-          if (isYear1AfterYear2(year, currentYear)) {
-            console.log("AHEAD of TIME!");
-          } else {
-            if (isYear1BeforeYear2(year, "2020")) {
-              setFootballStandings(footballStandings);
-            }
-            display(trifectaStandings, h2hStandings, top5Bottom5Standings);
-          }
-          return;
-        }
-
-        // if there is no last scraped string (ie brand new, first time entering), scrape
-        if (!lastScrapedString) {
-          console.log("SHOULD SCRAPE BUT DO NOT FOR TESTING");
-          scrape(collection);
+        if (isYear1BeforeYear2(year, "2020")) {
+          setFootballStandings(footballStandings);
         } else {
-          const now = new Date();
-          const alreadyScraped = isSameDay(now, new Date(lastScrapedString));
-
-          // only scrape if not already scraped today
-          if (!alreadyScraped) {
-            console.log("SHOULD SCRAPE BUT DO NOT FOR TESTING");
-            scrape(collection);
-          } else {
-            console.log("Already scraped");
-            display(trifectaStandings, h2hStandings, top5Bottom5Standings);
-          }
+          display(trifectaStandings, h2hStandings, top5Bottom5Standings);
         }
       };