+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ### PR links (https://github.com/wesleychipeco/trifecta-website-v3/pulls?q=is%3Apr+is%3Aclosed)
 
+### [4.3.0] - 2026-06-12
+
+### [PR-105] - 2026-06-12
+
+- Change MongoDB connection library and process
+- Add 2026 Baseball draft results
+
 ### [PR-104] - 2025-12-30
 
 - Update Fantrax API version for requests